Initialize ASP.NET Web API and serve original application as static files
We still use MVC in the application, but the view part will do nothing.
It's important to serve the web application as static files since we
don't want to add complexity to the code like data-binding. Plus, it
would be counter-intuitive to deviate away from using pure, vanilla
JavaScript code when the whole purpose of branching away to this version
was to provide simple and maintainable code to the staffs of Office of
Elections.